+++ R/warming_stripes_plot.R
@@ -3,6 +3,8 @@
 #' This plotting routine generates graphical output for
 #' the given variable within the given time range and area.
 #' Dependent on the output format a PNG is created.
+#' 
+#' Circular stripe plots are inspired by Emanuele Bevacqua (see emanuele.bevacqua.eu)
 #'
 #' @inheritParams monitor_climate
 #' @param pointsTF Show data points (logical).
@@ -15,6 +17,7 @@
 warming_stripes_plot <- function(variable = NULL,
                                  infile = NULL, 
                                  selected_number = 1,
+                                 stripe_color = 1,
                                  analyze_method = TRUE,
                                  temp_dir = tempdir(),
                                  out_dir = getwd(),
@@ -36,7 +39,8 @@ warming_stripes_plot <- function(variable = NULL,
                                  attach = FALSE,
                                  infile_attach = "auto",
                                  pointsTF = FALSE,
-                                 lineTF = FALSE, 
+                                 lineTF = FALSE,
+                                 circ_plot = FALSE,
                                  title = "",
                                  verbose = TRUE,
                                  nc = NULL)
@@ -176,7 +180,9 @@ warming_stripes_plot <- function(variable = NULL,
   }
   
   country_name <- get_country_name(country_code)
-  title <- paste0(variable, " Stripes Plot: ", country_name, " (", climate_year_start, " - ", substring(end_date,1,4), ")")
+  # title <- paste0(variable, " Stripes Plot: ", country_name, " (", climate_year_start, " - ", substring(end_date,1,4), ")")
+  title <- paste0(variable, " change in ", country_name, " since XXXX (Reference: ", 
+                  climate_year_start, " - ", substring(end_date,1,4), ")")
   
   plot_warming_stripes(
     variable = variable,
@@ -193,5 +199,7 @@ warming_stripes_plot <- function(variable = NULL,
     pointsTF = pointsTF,
     lineTF = lineTF, 
     title = title,
+    stripe_color = stripe_color,
+    circ_plot = circ_plot,
     verbose = TRUE)
 }